+++ docs/Changelog.md
@@ -4,6 +4,7 @@
 ## v3.5 (upcoming)
 - FIX: Display Performance Report Per Test Case
 - FIX: NullPointerException in ConstraintFactory
+- FIX: NumberFormatException in RelativeConstraint
 - FIX: RelativeConstraint.tolerance and AbsoluteConstraint.value fields in UI
 - ADD: move some options to Advanced
 - ADD: choosing display percentiles